new try

+++ hashmap.c
@@ -81,7 +81,7 @@ HashMap * createMap(long capacity)
 
 void eraseMap(HashMap * map,  char * key) {    
 
-  for (int i = 0; i < map->capacity-1; i++)
+  for (long i = 0; i < map->capacity-1; i++)
     {
       if(map->buckets->key == key)
       {
